# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Got Journey Thailand API",
    "description": "Travel Booking Platform API - Hotels, Cars, and Package Bookings",
    "version": "1.0.0",
    "contact": {
      "name": "Got Journey Thailand",
      "email": "support@gotjourneythailand.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    },
    {
      "url": "https://api.gotjourneythailand.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/cars/{id}": {
      "get": {
        "tags": ["Cars"],
        "summary": "Get car by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" },
          "404": { "description": "Car not found" }
        }
      },
      "put": {
        "tags": ["Cars"],
        "summary": "Update car (Admin only)",
        "security": [{ "AdminAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Car updated" }
        }
      },
      "delete": {
        "tags": ["Cars"],
        "summary": "Delete car (Admin only)",
        "security": [{ "AdminAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Car deleted" }
        }
      }
    },
    "/api/bookings": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get all bookings",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      },
      "post": {
        "tags": ["Bookings"],
        "summary": "Create a new booking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookingInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Booking created" }
        }
      }
    },
    "/api/bookings/{code}": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get booking by code",
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" },
          "404": { "description": "Booking not found" }
        }
      },
      "put": {
        "tags": ["Bookings"],
        "summary": "Update booking",
        "parameters": [
          { "name": "code", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Booking updated" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      },
      "UserAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    },
    "schemas": {
...
      "BookingInput": {
        "type": "object",
        "required": ["booking_type", "check_in_date", "check_out_date", "customer_name", "customer_email", "customer_phone"],
        "properties": {
          "booking_type": { "type": "string", "enum": ["HOTEL", "CAR", "COMBO"] },
          "hotel_id": { "type": "string" },
          "car_id": { "type": "string" },
          "check_in_date": { "type": "string", "format": "date" },
          "check_out_date": { "type": "string", "format": "date" },
          "number_of_guests": { "type": "integer" },
          "customer_name": { "type": "string" },
          "customer_email": { "type": "string", "format": "email" },
          "customer_phone": { "type": "string" },
          "customer_line": { "type": "string" },
          "special_requests": { "type": "string" }
        }
      },
...
    }
  }
}